update task 11

+++ 0x04-more_functions_nested_loops/main.h
@@ -31,6 +31,6 @@ void FizzBuzz(void);
 
 void print_triangle(int size);
 
-int isprime(int n, int firsttime);
+int isprime(long int n, int firsttime);
 
 #endif /* MAIN_H */
